pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "committee-history-builder"
version = "0.1.0"
description = "Python-first CLI for generating standalone committee history HTML pages from YAML."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "Jinja2>=3.1.4",
  "markdown-it-py>=3.0.0",
  "mdit-py-plugins>=0.4.1"
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0"]

[project.scripts]
committee = "committee_builder.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
